Add files via upload
This is a Data set that collected 10 individuals responses to the following questions. 

Q1: There is widespread mental health epidemics, problems such as depression and anxiety and many more. But WHY? In our opinion, what is the leading cause and what is the second-leading cause of this collective increase? For each of them can you briefly mention how that factor leads to a mental health problem?

Q2: What is your favorite thing about the modern human life, what is your second-favorite thing, and what is your least-favorite thing?

Q3: name the biggest factor that, in our opinion, today, harms the mental health of mothers specifically.

Q4: name the biggest factor that, in our opinion, today, harms the mental health of fathers specifically. 

Q5: name the biggest factor that, in our opinion, today, harms the mental health of children specifically.